added mouseX and translate
Forgot to add mouseX so that the sprite would follow the mouse along the x access. Also had to add a translate function so that the sprite would go all the way to the left edge of the canvas. Sandwiched it in between push and pop so it (hopefully) won't affect the rest of my code later on.